+++ Projects/DeferredPBR/SceneDeferredPBR.cc
@@ -23,16 +23,17 @@ void SceneDeferredPBR::OnInit() {
   glEnable(GL_DEPTH_TEST);
   view_ = glm::lookAt(glm::vec3(0.0f, 3.0f, 5.0f), glm::vec3(0.0f, 0.0f, 0.0f),
                       glm::vec3(0.0f, 1.0f, 0.0f));
-  proj_ = glm::perspective(glm::radians(kFOVY),
-                           static_cast<float>(width_) / height_, 0.1f, 100.0f);
+  proj_ = glm::perspective(
+      glm::radians(kFOVY),
+      static_cast<float>(width_) / static_cast<float>(height_), 0.1f, 100.0f);
 
   CreateVAO();
 
   gbuffer_.OnInit(width_, height_);
 
   lightPositions_ = {glm::vec4(7.0f, 3.0f, 0.0f, 1.0f),
-                      glm::vec4(0.0f, -0.15f, 0.0f, 0.0f),
-                      glm::vec4(-8.0f, 3.0f, 0.0f, 1.0f)};
+                     glm::vec4(0.0f, -0.15f, 0.0f, 0.0f),
+                     glm::vec4(-8.0f, 3.0f, 0.0f, 1.0f)};
 
   prog_.SetUniform("Light[0].L", glm::vec3(45.0f));
   prog_.SetUniform("Light[0].Position", view_ * lightPositions_[0]);
@@ -72,7 +73,8 @@ void SceneDeferredPBR::OnRender() {
 void SceneDeferredPBR::OnResize(int w, int h) {
   SetDimensions(w, h);
   glViewport(0, 0, w, h);
-  proj_ = glm::perspective(glm::radians(kFOVY), static_cast<float>(w) / h, 0.3f,
+  proj_ = glm::perspective(glm::radians(kFOVY),
+                           static_cast<float>(w) / static_cast<float>(h), 0.3f,
                            100.0f);
 }
 
@@ -144,8 +146,9 @@ void SceneDeferredPBR::Pass1() {
 
   view_ = glm::lookAt(glm::vec3(0.0f, 4.0f, 7.0f), glm::vec3(0.0f, 0.0f, 0.0f),
                       glm::vec3(0.0f, 1.0f, 0.0f));
-  proj_ = glm::perspective(glm::radians(kFOVY),
-                           static_cast<float>(width_) / height_, 0.3f, 100.0f);
+  proj_ = glm::perspective(
+      glm::radians(kFOVY),
+      static_cast<float>(width_) / static_cast<float>(height_), 0.3f, 100.0f);
   DrawScene();
 }
 
@@ -175,8 +178,8 @@ void SceneDeferredPBR::DrawScene() {
   DrawFloor();
 
   for (int i = 0; i < kNumCows; i++) {
-    const float cowX = i * (20.0f / (kNumCows - 1)) - 10.0f;
-    const float rough = (i + 1) * (1.0f / kNumCows);
+    const float cowX = static_cast<float>(i) * (20.0f / (kNumCows - 1)) - 10.0f;
+    const float rough = static_cast<float>(i + 1) * (1.0f / kNumCows);
     DrawMesh(glm::vec3(cowX, 0.0f, 0.0f), rough, 0, kBaseCowColor);
   }
